+++ lib/ts/recipe/emailverification/utils.ts
@@ -19,7 +19,7 @@ import { normaliseRecipeModuleConfig } from "../recipeModule/utils";
 import { RecipeInterface } from "supertokens-web-js/recipe/emailverification";
 
 export function normaliseEmailVerificationFeature(config: Config): NormalisedConfig {
-    const disableDefaultImplementation = config.disableDefaultImplementation === true;
+    const disableDefaultUI = config.disableDefaultUI === true;
     const mode = config.mode === undefined ? "OFF" : config.mode;
 
     const sendVerifyEmailScreenStyle =
@@ -48,7 +48,7 @@ export function normaliseEmailVerificationFeature(config: Config): NormalisedCon
 
     return {
         ...normaliseRecipeModuleConfig(config),
-        disableDefaultImplementation,
+        disableDefaultUI,
         mode,
         sendVerifyEmailScreen,
         verifyEmailLinkClickedScreen,