+++ CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Adds an example app with emailpassword + vercel
 
+## [0.20.5] - 2022-05-04
+
+### Fixes
+
+-   Fixed support for react 18 strict mode
+
+### Changes
+
+-   Updated react dependency to react 18
+
 ## [0.20.4] - 2022-04-11
 
 -   Makes error message for changing of props in `SessionAuth` more clear.